requirements.txt
opencv-python
pytesseract
pandas
openpyxl
pillow
argparse
requests